requirements.txt
uvicorn==0.12.1
fastapi==0.61.1
paho-mqtt==1.5.0
matplotlib==3.3.4
numpy==1.18.1